Do not use JUL during launcher startup, defer logging to the applicatin phase
To allow Spring Boot applications to redirect JUL to the logging sub-system
which has been configured for the application, the launcher must not use
Java Util Logging. As soon as you invoke a log method on the JUL's Logger
class, the LogManager will be initialized and that cannot be changed
later on. That means that previously any attempt in the application phase
to redirect Java Util Logging did not succeed, resulting in mixed stdout
formats, one from the application's logging sub-system and one from JUL.

This changeset collects all log messages during the launcher phase
and transfers it to the application phase so that it can be replayed
as soon as the application's logging sub-system has been initialized.

Some of the design choices have been doumented as part of the
DeferredLogger class' Javadoc. Fixes gh-3815 (partially)